edu modal started; edu-osc set; +hbs for edu modal
