+++ index.html
@@ -11,7 +11,8 @@
 <body>
     <div id="nav">
     </div>
-    <div id="eduModal"></div>
+    <div id="eduModal">
+    </div>
     <div id="editModal">
     </div>
     <div id="saveModal">
@@ -144,21 +145,15 @@
         <input type="radio" name="patch" id="patch4">
         </div>
     </div>
-    <div>
-        <canvas id="oscScope"></canvas>
-        <button id="startSine">start sine</button>
-        <button id="startSquare">start square</button>
-        <button id="startTriangle">start triangle</button>
-        <button id="startSawtooth">start sawtooth</button>
-    </div>
+
     <div id="chartWrap">
         <canvas id="adsr"></canvas>
     </div>
     <div id="chartCtrl">
-    <input type="range" name="attack" id="attack" step="0.01" min="0" max="1" value=".4">
-    <input type="range" name="decay" id="decay" step="0.01" min="0" max="1" value=".8">
-    <input type="range" name="sustain" id="sustain" step="0.01" min="0" max="1" value=".6">
-    <input type="range" name="release" id="release" step="0.01" min="0" max="1" value=".1">
+        <input type="range" name="attack" id="attack" step="0.01" min="0" max="1" value=".4">
+        <input type="range" name="decay" id="decay" step="0.01" min="0" max="1" value=".8">
+        <input type="range" name="sustain" id="sustain" step="0.01" min="0" max="1" value=".6">
+        <input type="range" name="release" id="release" step="0.01" min="0" max="1" value=".1">
     </div>
     <!-- <h3>&darr; qwerty keys &darr;</h3>
     <img id="piano" src="keys.jpg"> -->
