+++ sass/main.scss
@@ -15,6 +15,24 @@ $keyBorder: .3vw double white;
     background-color: rgba(61, 61, 61, 0.836);
 }
 
+@mixin closeChip {
+    position: absolute;
+    display: inline-block;
+    left: 90%;
+    top: 10%;
+    font-size: 3vw;
+    color: rgb(49, 49, 49);
+    border-radius: 50%;
+    &:hover{
+        color: rgb(2, 2, 122);
+        cursor: pointer;
+        &:active {
+            color: gold;
+            text-shadow: 0 0 2vw rgb(82, 81, 81);
+        }
+    }
+}
+
 * {
     margin: 0;
 }
@@ -71,6 +89,44 @@ $keyBorder: .3vw double white;
 }
 
 //modals
+
+#eduModal {
+    @include modal;
+    #eduWrap{
+        position: relative;
+        margin: auto;
+        padding: 2vw;
+        width: 82%;
+        height: auto;
+        border-radius: 3em;
+        color: rgb(51, 51, 51);
+        background: white;
+        #oscView{
+            main{
+                float: left;
+                width: 40%;
+                text-align: center;
+            } 
+            aside {
+                float: right;
+                width: 55%;
+                div {
+                    margin: .8vw;
+                }
+            }
+            #pickOsc{
+                clear: both;
+            }
+        }
+        .closeChip {
+            @include closeChip;
+            left: 95%;
+            top: 5%;
+        }
+    }
+}
+
+
 #editModal, #saveModal, #deleteModal {
     @include modal;
     .patchModal{
