+++ templates/in-nav.hbs
@@ -7,3 +7,4 @@
             {{/each}}
     </div>
 </div>
+<div id="startBuild"><a>start synthbuilder</a></div>
