+++ templates/osc.hbs
@@ -1,0 +1,28 @@
+<div id="eduWrap">
+<div id="oscView">
+    <main>
+        <h3>Oscillator</h3>
+        <p>the core to any synthesizer</p>
+    </main>
+    <aside>
+        <div>
+        <button id="startSine">start sine</button>
+        <img src="../images/sine.png">
+        </div>
+        <div>
+        <button id="startTriangle">start triangle</button>
+        <img src="../images/triangle.png">
+        </div>
+        <div>
+        <button id="startSquare">start square</button>
+        <img src="../images/square.png">
+        </div>
+        <div>
+        <button id="startSawtooth">start sawtooth</button>
+        <img src="../images/sawtooth.png">
+        </div>
+    </aside>
+    <div id="pickOsc"></div>
+</div>
+    <span class="closeChip">&times;</span>
+</div>
